lukefile: external_dependencies only works at the top-level with a library key.
This is a subset of the functionality provided by LuaRocks proper, because it
currently supports only a `library={<checksymbol-config>}` and does not perform
library or header-file search.
* lib/luke/lukefile.lua (unwind_external_dependencies): Rename from this...
(unwrap_external_dependencies): ...to this. Adjust all callers.
Remove recursion loop, we only support external_dependencies at the top level.
Don't add '-l' prefix to library name, because it was added already during
run_configs for checksymbol config.
* build-aux/luke: Regenerate

Signed-off-by: Gary V. Vaughan <gary@gnu.org>